Initial commit
Made a C++ SFML application with a race track and "cars" that move
around the track. The cars are spawned on a different thread.

The vector with moving cars and a mutex for locking access to this
vector are owned by the main function. Previously, when this function
terminated, the vector and mutex were freed, while spawning thread still
operated on it, which resulted in a Segmentation Fault.

To fix this, `std::shared_ptr` was used, which counts the references to
itself and frees its contents only when there are no more references
left.

In our program, after creating the spawning thread there are 2
references for each pointer. After main returns, it starts cleaning up
its locals, by running its destructors. The destructors of the pointers
decrease the refcount so that only 1 reference remains in the spawning
thread. Next, the thread is stopped and joined, hence now the thread
proceeds to freeing its locals. The destructors of the pointers are
decremented to 0, and so the entire pointer is dealocated by the
spawning thread.

The destructor of the thread returns successfully and the program closes
without any errors. I only knew how to fix this because I encountered
this same issue in Rust.